Split cids access approve request scenarios into separate spec file
